+++ CMakeLists.txt
@@ -31,7 +31,7 @@ set(OPENING DeceideOverOpening)
 set(BI BoardImplementation)
 set(FAC FactorialImplementation)
 set(SF StockfishSource)
-set(LOGGER CustomLogger)
+set(UTILS GeneralUtilities)
 set(WRAP EngineWrapper)
 set(ENGINE SwitchEngine)
 set(DFS1P DepthFirstSearchOnePerson)
