+++ app/CMakeLists.txt
@@ -20,6 +20,14 @@ set(EXE_HEADERS
     "SwitchMain.h")
 
 add_compile_definitions(CMAKE_SOURCE_DIR="${CMAKE_SOURCE_DIR}")
+message("Build type: " ${CMAKE_BUILD_TYPE})
+if(CMAKE_BUILD_TYPE MATCHES Debug)
+    add_compile_definitions(DEBUG)
+    add_compile_definitions(MODE_CLI=1)
+else(CMAKE_BUILD_TYPE MATCHES Debug)
+    add_compile_definitions(MODE_CLI=0)
+endif(CMAKE_BUILD_TYPE MATCHES Debug)
+    
 add_executable(${EXECUTABLE} 
     ${EXE_SOURCES}
     ${EXE_HEADERS})
