+++ app/SwitchMain.cpp
@@ -1,12 +1,4 @@
-// #include "config.hpp"
 #include "SwitchMain.h"
-#ifdef _WIN32
-#include <direct.h>
-#define GETCWD _getcwd
-#else
-#include <unistd.h>
-#define GETCWD getcwd
-#endif
 
 // https://stackoverflow.com/questions/11826554/standard-no-op-output-stream/11826666#11826666
 class NullBuffer : public std::streambuf {
@@ -15,9 +7,6 @@ public:
 };
 NullBuffer null_buffer;
 std::ostream null_stream(&null_buffer);
-
-#define DEBUG
-#define MODE_CLI 1
 
 #ifdef DEBUG
 #define debug std::cerr
@@ -38,31 +27,6 @@ void chessGameLoop(SwitchEngine &engine) {
 		engine.processMove(opp);
 	}
 }
-
-/**
- * @brief Convert int on 64 to algebraic notation
- *
- * @param pos postition
- * @return std::string algebraic notation of the position
- */
-std::string toAN(int pos) {
-	int row = pos >> 3;
-	int col = pos & 7;
-	return std::string({(char)col + 'a', 7 - (char)row + '1'});
-}
-
-/**
- * @brief Convert algebraic notation to int on 64
- *
- * @param str algebraic notation
- * @return int position (8 * row + col)
- */
-int parseAN_(std::string str) {
-	int row = 7 - (str[1] - '1');
-	int col = str[0] - 'a';
-	return row * 8 + col;
-}
-// @TODO: refactor to remove code duplication with `SwitchEngine.cpp`
 
 void CLIGameLoop(SwitchEngine &engine) {
 	while (true) {
@@ -85,8 +49,8 @@ void CLIGameLoop(SwitchEngine &engine) {
 				opponentMove[1] <= '8' && '1' <= opponentMove[3] &&
 				opponentMove[3] <= '8') {
 			Closedfish::Move opp =
-					std::make_tuple(parseAN_(opponentMove.substr(0, 2)),
-													parseAN_(opponentMove.substr(2, 2)), 0.0);
+					std::make_tuple(parseAN(opponentMove.substr(0, 2)),
+													parseAN(opponentMove.substr(2, 2)), 0.0);
 			debug << "[DEBUG] " << std::get<0>(opp) << " " << std::get<1>(opp)
 						<< std::endl;
 			engine.processMove(opp);
@@ -104,22 +68,21 @@ void opt(const Stockfish::UCI::Option &o) { Stockfish::Eval::NNUE::init(); };
 
 int main(int argc, char *argv[]) {
 	Closedfish::Logger logger;
-	debug << Stockfish::engine_info() << std::endl;
-	char buff[40000];
-	char *cwd = GETCWD(buff, 40000);
-	std::filesystem::path p = std::filesystem::path(CMAKE_SOURCE_DIR) /
-														"external" / "stockfish" / "src" /
-														"nn-ad9b42354671.nnue";
-	/*
-	std::ifstream stream(Stockfish::Options["EvalFile"], std::ios::binary);
-	if (stream.fail()) {
-		debug << "EvalFile doesn't exist. Check NNUE EvalFile." << std::endl;
-		exit(-1);
-	}
-	stream.close();
-	*/
+	debug << "[INFO] " << Stockfish::engine_info() << std::endl;
 	Stockfish::CommandLine::init(argc, argv);
 	Stockfish::UCI::init(Stockfish::Options);
+	std::filesystem::path p = std::filesystem::path(CMAKE_SOURCE_DIR) /
+														"external" / "stockfish" / "src" /
+														(std::string)Stockfish::Options["EvalFile"];
+	std::ifstream stream(p.string(), std::ios::binary);
+	if (stream.fail()) {
+		debug << "[ERROR] EvalFile doesn't exist. Expected directory: "
+					<< p.string() << std::endl;
+		exit(-1);
+	} else {
+		debug << "[INFO] EvalFile found." << std::endl;
+	}
+	stream.close();
 	Stockfish::Options["EvalFile"]
 			<< Stockfish::UCI::Option(p.string().c_str(), opt);
 	Stockfish::Tune::init();
