+++ lib/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(engine_wrapper)
 add_subdirectory(board_implementation)
-add_subdirectory(logger)
+add_subdirectory(utils)
 add_subdirectory(DFS1P)
 add_subdirectory(weak_pawns)
 add_subdirectory(heatmap)
