+++ lib/board_implementation/CFBoard.h
@@ -1,13 +1,12 @@
 #pragma once
 
 #ifdef _MSC_VER
-#include <nmmintrin.h>
 #include <immintrin.h>
+#include <nmmintrin.h>
 #define __builtin_popcountll _mm_popcnt_u64
 #define __builtin_ctzll _tzcnt_u64
 #define __builtin_clzll _lzcnt_u64
 #endif
-
 #include <iostream>
 #include <stdint.h>
