+++ src/connectors/StockfishConnect/CMakeLists.txt
@@ -11,7 +11,7 @@ target_include_directories(${SC} PUBLIC
     "./"
     "${CMAKE_BINARY_DIR}/configured_files/include")
 
-target_link_libraries(${SC} PUBLIC ${BI} ${WRAP} ${SF} ${LOGGER})
+target_link_libraries(${SC} PUBLIC ${BI} ${WRAP} ${SF} ${UTILS})
 
 if (${ENABLE_WARNINGS})
     target_set_warnings(TARGET ${SC} ENABLE ON AS_ERROR OFF)
