+++ src/connectors/StockfishConnect/StockfishConnect.h
@@ -1,16 +1,7 @@
 #pragma once
 #include <EngineWrapper.h>
-#include <logger.h>
-#include <stockfish/src/bitboard.h>
-#include <stockfish/src/endgame.h>
-#include <stockfish/src/position.h>
-#include <stockfish/src/psqt.h>
-#include <stockfish/src/search.h>
-#include <stockfish/src/syzygy/tbprobe.h>
-#include <stockfish/src/thread.h>
-#include <stockfish/src/tt.h>
-#include <stockfish/src/uci.h>
 #include <tuple>
+#include <utils.h>
 
 class StockfishEngine : public Closedfish::ChessEngine {
 public:
